pyproject.toml
[tool.poetry]
name = "aibetrade-stock"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
telethon = "^1.34.0"
python-dotenv = "^1.0.1"
openai = "^1.14.3"
langchain = "^0.1.13"
loguru = "^0.7.2"
flask = "^3.0.3"
redis = "^5.0.3"
binance = "^0.3"
python-binance = "^1.0.19"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.30"
werkzeug = "^3.0.3"
aiogram = "^3.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
